backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
psycopg[binary]==3.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0

